Rename package for TypeAdapters
