+++ src/main/java/pl/jkuznik/utils/persistentState/StateReader.java
@@ -14,9 +14,9 @@ import pl.jkuznik.computer.hardware.shared.Component;
 import pl.jkuznik.computer.hardware.shared.FileHandler;
 import pl.jkuznik.computer.hardware.shared.enums.ComponentType;
 import pl.jkuznik.computer.software.file.File;
-import pl.jkuznik.utils.persistentState.gson.ComponentGsonAdapter;
-import pl.jkuznik.utils.persistentState.gson.FileAdapterGson;
-import pl.jkuznik.utils.persistentState.gson.FileHandlerAdapterGson;
+import pl.jkuznik.utils.persistentState.gsonTypeAdapter.ComponentGsonAdapter;
+import pl.jkuznik.utils.persistentState.gsonTypeAdapter.FileAdapterGson;
+import pl.jkuznik.utils.persistentState.gsonTypeAdapter.FileHandlerAdapterGson;
 
 import java.io.IOException;
 import java.nio.file.Files;
